+++ setupcode/setupcode.go
@@ -13,14 +13,14 @@ import (
 // References are:
 // - https://github.com/nfarina/homebridge/issues/1588#issuecomment-341158722
 // - https://github.com/KhaosT/HAP-NodeJS/blob/67032e75b9f5f74993ad932c849d5bbb1937a097/lib/Accessory.js#L363
-func GenXhmURI(category uint, hapType uint, pin string, setupID string) string {
+func GenXhmURI(category int64, hapType int64, pin string, setupID string) string {
 
 	prefix := "X-HM://00"
 	var payload int64
 
 	payload = 0
 	cat := category << 31
-	payload |= int64(cat)
+	payload |= cat
 	logrus.Debugf("payload with cat: %d", payload)
 
 	ip := 1 << 28
